pyproject.toml
[tool.poetry]
name = "image-jinn"
version = "1.0.0"
description = "High-performance image storage service"
authors = ["Mohammed Rishad <mohammedrishad.vk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.13"
djangorestframework = "3.13.1"
drf-spectacular = "0.22.1"
django-cors-headers = "3.11.0"
psycopg2-binary = "2.9.3"
boto3 = "1.24.7"
celery = "5.2.7"
redis = "4.3.3"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
# Make sure this matches `*.py` in .editorconfig
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
